requirements.txt
flask
marshmallow
python-abc
python-dotenv
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
requests==2.32.3
flask-jwt-extended==4.4.4